chore: move hub-icon to components root
