+++ src/components/unread-dot/index.tsx
@@ -11,8 +11,7 @@ type Props = {
  * Notification icon unread dot UI component.
  *
  * @param props
- * @param props.isActive Predicate of whether the unread dot is in an
- *                       active state.
+ * @param props.isActive Predicate of whether the unread dot is in an active state.
  */
 export default function UnreadDot( { isActive }: Props ) {
 	return (
