+++ src/wp-notifications.scss
@@ -8,7 +8,7 @@
 /* Components */
 @import "styles/components/notification";
 @import "styles/components/bullet";
-@import "components/unread-dot/style";
+// @import "components/unread-dot/style";
 
 // Dashboard notifications
 @import "styles/dashboard/notice";
